backend/requirements.txt
flask==3.0.0
torch==2.0.1
transformers==4.33.2
pillow==10.0.1
python-dotenv==1.0.0